+++ src/windows/windisplaydevicegeneral.cpp
@@ -59,12 +59,12 @@ namespace display_device {
         }
 
         const auto display_name { m_w_api->getDisplayName(best_path) };
-        const float refresh_rate { best_path.targetInfo.refreshRate.Denominator > 0 ?
-                                     static_cast<float>(best_path.targetInfo.refreshRate.Numerator) / static_cast<float>(best_path.targetInfo.refreshRate.Denominator) :
-                                     0.f };
+        const double refresh_rate { best_path.targetInfo.refreshRate.Denominator > 0 ?
+                                      static_cast<double>(best_path.targetInfo.refreshRate.Numerator) / static_cast<double>(best_path.targetInfo.refreshRate.Denominator) :
+                                      0. };
         const EnumeratedDevice::Info info {
           { source_mode->width, source_mode->height },
-          m_w_api->getDisplayScale(display_name, *source_mode).value_or(0.f),
+          m_w_api->getDisplayScale(display_name, *source_mode).value_or(0.),
           refresh_rate,
           win_utils::isPrimary(*source_mode),
           { source_mode->position.x, source_mode->position.y },
